rearchitecure motivation view to properly use protocols
