+++ Productivity Tracker.xcworkspace/xcuserdata/matejbutkovic.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "B57C29A0-B97D-440C-B80C-35A3FAD04B12"
+   type = "0"
+   version = "2.0">
+</Bucket>
